+++ project/src/test/mailbox/writer2 (1).c
@@ -18,7 +18,7 @@ Siccome le mailbox persistono anche quando i processi muoiono, per evitare che u
 #define KEYFILE "progfile"
 
 typedef struct msg {
-    long mesg_type;
+    int mesg_type;
     char text[MAXMSG];
     short int valore;
     time_t session;
